chore: rename semantic tokens file
