+++ src/lib/panda/index.ts
@@ -6,5 +6,5 @@ export { default as fonts } from "./fonts";
 export { default as globalCss } from "./globalCss";
 export { default as keyframes } from "./keyframes";
 export { recipes, slotRecipes } from "./recipes";
-export { default as semanticTokens } from "./semantic-tokens";
+export { default as semanticTokens } from "./semanticTokens";
 export { default as zIndex } from "./z-index";
